Add menu items to the hamburger menu (mobile).
In order to restrict the implementation to pure HTML/CSS, the
interaction with the menu is emulated using a checkbox which will toggle
the menu visible or invisible.
The menu itself is a list of links.

NOTE: the menu is functional, but not yet styled; i.e., this is a WIP.