Implement contact form
Issue: because the input placeholder contains an image and text, it is
not possible  -- at least not in a simple way -- to implement the
gradient blue color specified in the Figma.
Thus, I opted for using the darkest color from the gradient for the
placeholder text.
Note, however, tha the input border has the gradient color as specified.

Almost done, but I cannot make the mail icon show.
The problem is related with how I implement the degradé font color in
the placeholder.